requirements.txt

requests
pandas
pyodbc
python-dotenv